+++ roxctl/sensor/sensor.go
@@ -4,6 +4,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/stackrox/rox/roxctl/common/flags"
 	"github.com/stackrox/rox/roxctl/sensor/generate"
+	"github.com/stackrox/rox/roxctl/sensor/getbundle"
 )
 
 // Command controls all of the functions being applied to a sensor
@@ -13,7 +14,10 @@ func Command() *cobra.Command {
 		Short: "Sensor is the list of commands that pertain to the Sensor service",
 		Long:  "Sensor is the list of commands that pertain to the Sensor service",
 	}
-	c.AddCommand(generate.Command())
+	c.AddCommand(
+		generate.Command(),
+		getbundle.Command(),
+	)
 	flags.AddTimeout(c)
 	return c
 }